在卡片上添加显示神婆注册日期，固定格式，分三个板块，new挑选出两个最新注册的神婆，live挑选出剩余的中在线的神婆4个，按注册时间最老的开始排，pick再从剩余的神婆中挑选6个，按注册时间从最老的开始排

+++ flutter_app/lib/models/star.dart
@@ -33,7 +33,7 @@ class Star {
   @JsonKey(defaultValue: 0)
   int workStatus;
   @JsonKey(defaultValue: 0.0)
-  double firstApprovedAt;
+  int firstApprovedAt;
   @JsonKey(defaultValue: '')
   String featureCover;
